+++ lib/distributed_job.rb
@@ -82,15 +82,13 @@ class DistributedJob
   #   end
 
   def push_each(enum)
-    return enum_for(:push_each, enum) unless block_given?
-
     previous_object = nil
     previous_index = nil
 
     enum.each_with_index do |current_object, current_index|
       push(current_index)
 
-      yield(previous_object, previous_index.to_s) if previous_index
+      yield(previous_object, previous_index.to_s) if block_given? && previous_index
 
       previous_object = current_object
       previous_index = current_index
@@ -98,14 +96,14 @@ class DistributedJob
 
     close
 
-    yield(previous_object, previous_index.to_s) if previous_index
+    yield(previous_object, previous_index.to_s) if block_given? && previous_index
   end
 
-  # Returns all pushed parts of the distributed job
+  # Returns all parts of the distributed job which are not yet finished.
   #
   # @return [Enumerator] The enum which allows to iterate all parts
 
-  def parts
+  def open_parts
     redis.sscan_each("#{redis_key}:parts")
   end