+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v2.0.0
+
+* `#push_each` no longer returns an enum when no block is given
+* Renamed `#parts` to `#open_parts`
+
 ## v1.0.0
 
 * Initial release